+++ lib/src/services/api_service.dart
@@ -1,7 +1,15 @@
 // Dart imports:
 import 'dart:async';
+import 'dart:math';
+import 'dart:typed_data';
 
 // Package imports:
+import 'package:archethic_lib_dart/src/crypto.dart' as crypto;
+import 'package:archethic_lib_dart/src/model/crypto/key_pair.dart';
+import 'package:archethic_lib_dart/src/model/response/transaction_inputs_response.dart';
+import 'package:archethic_lib_dart/src/model/transaction_input.dart';
+import 'package:archethic_lib_dart/src/services/address_service.dart';
+import 'package:archethic_lib_dart/src/utils.dart' as utils;
 import 'package:http/http.dart' as http show Response, post;
 import 'package:logger/logger.dart';
 
@@ -332,4 +340,79 @@ class ApiService {
     _completer.complete(transactionsList);
     return _completer.future;
   }
+
+  /// Query the network to list the transaction inputs from an address
+  /// @param {String} The address scalar type represents a cryptographic hash used in the ArchEthic network with an identification byte to specify from which algorithm the hash was generated. The Hash appears in a JSON response as Base16 formatted string. The parsed hash will be converted to a binary and any invalid hash with an invalid algorithm or invalid size will be rejected
+  Future<List<TransactionInput>> getTransactionInputs(String address) async {
+    final Completer<List<TransactionInput>> _completer = Completer<List<TransactionInput>>();
+    List<TransactionInput> transactionInputs = List<TransactionInput>.empty(growable: true);
+    TransactionInputsResponse transactionInputsResponse = TransactionInputsResponse();
+    final Map<String, String> requestHeaders = {
+      'Content-type': 'application/json',
+      'Accept': 'application/json',
+    };
+    try {
+      final String _body =
+      'query { transactionInputs(address: \\"$address\\") { amount, from, nftAddress, spent, timestamp, type } }';
+      logger.d('getTransactionInputs: requestHttp.body=' + _body);
+      final http.Response responseHttp = await http.post(
+          Uri.parse(endpoint! + '/api'),
+          body: _body,
+          headers: requestHeaders);
+      logger.d('getTransactionInputs: responseHttp.body=' + responseHttp.body);
+      if (responseHttp.statusCode == 200) {
+        transactionInputsResponse =
+            transactionInputsResponseFromJson(responseHttp.body);
+            transactionInputs = transactionInputsResponse.data!.transactionInputs!;
+      }
+    } catch (e) {
+      logger.d('getTransactionInputs: error=' + e.toString());
+    }
+
+    _completer.complete(transactionInputs);
+    return _completer.future;
+  }
+
+      /// Create a keychain and an access keychain using the initial passphrase
+      /// @param {String | Uint8Array} passphrase Initial access passphrase
+      /// @param {*} originPrivateKey Origin private key
+      /// @returns Keychain transaction address
+      
+    /* void createKeychain(passphrase, originPrivateKey) async {
+         final Uint8List accessKeychainSeed = crypto.hash(passphrase);
+         final KeyPair publicKey = crypto.deriveKeyPair(utils.uint8ListToHex(accessKeychainSeed), 0);
+
+         final Uint8List keychainSeed = Uint8List.fromList(
+      List<int>.generate(12, (int i) => Random.secure().nextInt(32)));
+         final String keychainAddress = AddressService(endpoint).deriveAddress(utils.uint8ListToHex(keychainSeed), 0);
+
+         final Uint8List accessKeychainAesKey = Uint8List.fromList(
+      List<int>.generate(12, (int i) => Random.secure().nextInt(32)));
+
+Transaction accessKeychainTx = Transaction(
+                type: 'keychain_access', data: Transaction.initData())
+             .addSecret(crypto.aesEncrypt(keychainAddress, null), accessKeychainAesKey)
+             .addAuthorizedKey(publicKey, crypto.ecEncrypt(publicKey, accessKeychainAesKey))
+             .build(accessKeychainSeed, 0, 'P256')
+             .originSign(originPrivateKey);
+
+         KeyPair keyPair2 = crypto.deriveKeyPair(utils.uint8ListToHex(keychainSeed), 0);
+
+         final Uint8List keychainAesKey = Uint8List.fromList(
+      List<int>.generate(12, (int i) => Random.secure().nextInt(32)));
+
+         Transaction keyChainTx = Transaction( type: 'keychain', data: Transaction.initData())
+             .addSecret(crypto.aesEncrypt(JSON.stringify({ keychainSeed: keychainSeed }), keychain_aes_key))
+             .addAuthorizedKey(publicKey, crypto.ecEncrypt(keychain_aes_key), publicKey)
+             .addAuthorizedKey(keyPair2.publicKey, crypto.ecEncrypt(keychainAesKey), keyPair2.publicKey))
+             .build(keychainSeed, 0, 'P256')
+             .originSign(originPrivateKey);
+
+
+         const [access_keychain_res, keychain_res] = await Promise.all([this.sendTransaction(access_keychain_tx), this.sendTransaction(keychain_tx)])
+         if (access_keychain_res.status == "ok" && keychain_res.status == "ok") {
+             return keychain_tx.address
+         } else {
+             throw 'Something goes wrong !';
+         }*/
 }