+++ CHANGELOG.md
@@ -5,6 +5,7 @@ Changelog
 * Update dependencies.
 * Support multiple secrets.
 * Use integer instead of float
+* add getTransactionInputs method
 
 #### Version 1.0.0 (2021-08-08) 
 * Initial stable release.